fix encoding issues on python 3
Python 3 doesn't let you seek a file opened in "text" mode (the default,
where it's auto-decoded as utf-8) with a byte offset. This makes sense,
since otherwise you might jump into the middle of a utf-8 character.

This commit changes the stdout/stderr files to be opened in binary mode
so that the seeks to read the beginning/end of the stdout/stderr file
work and to manually decode as utf-8, throwing away any partial chars